# src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: E-Bookstore Platform API (BookWorm)
  description: >
    Complete Backend REST API specification for the E-Bookstore Capstone Project.
    Developed with AI-Augmented Workflow using IBM BOB.
    Supports user authentication, catalog browsing, recommendations, cart operations,
    shipping estimation, payment gateways (Cards, UPI, Wallets), gift point redemptions,
    and 48-hour order cancellation / returns.
  version: 1.0.0
  contact:
    name: Backend Engineering Team
    email: api-support@bookworm.example.com
servers:
  - url: http://localhost:8080/api/v1
    description: Local Development Server
...
paths:
...
  /payments/process:
    post:
      tags: [Payment & Wallet]
      summary: Process payment via Credit/Debit card, UPI, or Wallet
      operationId: processPayment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentRequest'
      responses:
        '200':
          description: Payment transaction status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentResponse'
  /payments/validate-coupon:
    post:
      tags: [Payment & Wallet]
      summary: Validate and apply discount coupon
      operationId: validateCoupon
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CouponRequest'
      responses:
        '200':
          description: Coupon calculation result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CouponResponse'
...
components:
  schemas:
...
    PaymentRequest:
      type: object
      required: [orderId, paymentMethod, amount]
      properties:
        orderId:
          type: integer
          example: 101
        paymentMethod:
          type: string
          enum: [CREDIT_CARD, DEBIT_CARD, UPI, WALLET]
          example: CREDIT_CARD
        cardNumber:
          type: string
          example: '4111-XXXX-XXXX-1111'
        cardHolderName:
          type: string
          example: Daniel Reed
        cvv:
          type: string
          example: '123'
        expiryDate:
          type: string
          example: '08/28'
        upiId:
          type: string
          example: daniel@okhdfcbank
        amount:
          type: number
          example: 470.00
    PaymentResponse:
      type: object
      properties:
        transactionId:
          type: string
          example: TXN-8947291834
        status:
          type: string
          example: SUCCESS
        paymentMethod:
          type: string
          example: CREDIT_CARD
        amountPaid:
          type: number
          example: 470.00
        timestamp:
          type: string
          format: date-time
    CouponRequest:
      type: object
      required: [couponCode, cartTotal]
      properties:
        couponCode:
          type: string
          example: SAVE100
        cartTotal:
          type: number
          example: 508.00
    CouponResponse:
      type: object
      properties:
        couponCode:
          type: string
          example: SAVE100
        discountAmount:
          type: number
          example: 100.00
        isValid:
          type: boolean
          example: true
        message:
          type: string
          example: Coupon applied successfully
...